Article Title should be H1
I get that H1 can also be the "Page Heading" aka Menu Title.

But in an article view, you're most likely not gonna wanna have the Menu Title as an H1. Or even if so, it's better to run the risk of having two H1 (if a menu name is displayed), than the article title not being an H1.

I personally don't care as I fix it in template overrides anyway. But as a standard, common practice, my two cents is that Article title should be H1 by default.